Add socket activation for systemd and s6
Use open, listening socket file descriptors  provided by the service manager
via LISTEN_FDS and LISTEN_PID environment variables.

This is how systemd tells the service about open sockets.
However, this change does not depend on systemd libraries or header.

All sockets must be mentioned exactly (IP:Port / unix-path)
in the mosquitto config file in order to be recognized.
Sockets not mentioned in the config file will be ignored.

The service/systemd directory contains an example (mosquitto.socket).
The service/s6 directory contains examples for the s6 service manager.

The new, builtin websocket mechanism can also use already open sockets.
However, the libwebsocket mchanism can not.

+++ src/listeners.c
@@ -251,7 +251,8 @@ void listeners__stop(void)
 		mosquitto__free(db.config->listeners[i].ws_protocol);
 #endif
 #ifdef WITH_UNIX_SOCKETS
-		if(db.config->listeners[i].unix_socket_path != NULL){
+		if(db.config->listeners[i].unix_socket_path != NULL &&
+		   db.config->listeners[i].unlink_on_close){
 			unlink(db.config->listeners[i].unix_socket_path);
 		}
 #endif
